# project/its-api-docs/spec/openapi.yaml
swagger: "2.0"
info:
  description: API for Issue tracking system
  version: 1.0.0
  title: ITS - API

# tags are used for organizing operations
...
paths:
...
  /project:
    post:
      tags:
        - project
      summary: create new project
      operationId: newProject
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/PROJECT_REQ"
      responses:
        200:
          description: new project success
          schema:
            $ref: "#/definitions/PROJECT_RES"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        409:
          description: repeated project name 
        500:
          description: Internal Server Error
    put:
      tags:
        - project
      summary: edit project detail
      operationId: editProject
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/EditProjectReq"
      responses:
        200:
          description: fetch profile success
          schema:
            $ref: "#/definitions/PROJECT_RES"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        409:
          description: repeated project name 
        500:
          description: Internal Server Error
  /project/get:
    post:
      tags:
        - project
      summary: fetch project
      operationId: fetchProject
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/GetProjectReq"
      responses:
        200:
          description: fetch project success
          schema:
            type: array
            items: 
              $ref: "#/definitions/PROJECT_RES"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        500:
          description: Internal Server Error
  /project/detail:
    post:
      tags:
        - project
      summary: fetch project detail
      operationId: fetchProjectDetail
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/ProjectDetailReq"
      responses:
        200:
          description: fetch detail success
          schema:
            type: array
            items: 
              $ref: "#/definitions/ProjectDetailRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        404:
          description: not found / permission denied
        500:
          description: Internal Server Error
  /project/tags:
    post:
      tags:
        - project
      summary: fetch project tags
      operationId: fetchProjectTags
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body 
          name: ProjectId
          schema:
            type: object
            required:
              - projectId
            properties:
              projectId:
                type: integer
                example: 1
      responses:
        200:
          description: fetch tags success
          schema:
            type: array
            items: 
              type : string
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        404:
          description: not found / permission denied
        500:
          description: Internal Server Error
  /project/member:
    post:
      tags:
        - project
      summary: add project member
      operationId: addProjectMember
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/ProjectMemberReq"
      responses:
        200:
          description: fetch profile success
          schema:
            $ref: "#/definitions/ProjectMemberRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        403:
          description: permission denied
        404:
          description: user not found 
        409:
          description: user already is a project member 
        500:
          description: Internal Server Error
    delete:
      tags:
        - project
      summary: remove project member
      operationId: removeProjectMember
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/DeleteProjectMemberReq"
      responses:
        200:
          description: remove success
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        403:
          description: permission denied
        404:
          description: user not found 
        500:
          description: Internal Server Error
    put:
      tags:
        - project
      summary: edit project member
      operationId: editProjectMember
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/UpdateProjectMemberReq"
      responses:
        200:
          description: update success
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        403:
          description: permission denied
        404:
          description: user not found 
        500:
          description: Internal Server Error
  /project/getMember:
    post:
      tags:
        - project
      summary: get project member by project id
      operationId: getProjectMember
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: body
          name: FormData
          schema:
            $ref: "#/definitions/GetProjectIssueReq"
      responses:
        200:
          description: fetch member success
          schema:
            type: array
            items: 
              $ref: "#/definitions/ProjectMemberRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        403:
          description: permission denied
        500:
          description: Internal Server Error
  /project/report:
    get:
      tags:
        - project
      summary: get project repoty
      operationId: getProjectReport
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          type: string
          name: sessionId
          required: true
        - in: query
          type: integer
          name: projectId
          required: true
        - in: query
          type: integer
          name: day
          required: true
        - in: query
          type: integer
          name: priority
          required: true
        - in: query
          type: boolean
          name: closed
          required: true
      responses:
        200:
          description: fetch report success
          schema:
            type: array
            items: 
              $ref: "#/definitions/ProjectReportRes"
        400:
          description: Invalid parameter
        401:
          description: sessionId unauthorized
        403:
          description: permission denied
        500:
          description: Internal Server Error
...
definitions:
...
  PROJECT_REQ:
    type: object
    required:
      - name
      - isPrivate
    properties:
      name:
        type: string
        example: ting's project
      description:
        type: string
        example: hello world
      isPrivate:
        type: boolean
        example: true
  EditProjectReq:
    type: object
    required:
      - name
      - isPrivate
      - id
      - lineId
    properties:
      name:
        type: string
        example: ting's project
      description:
        type: string
        example: hello world
      isPrivate:
        type: boolean
        example: true
      id:
       type: integer
      lineId:
       type: string
  GetProjectReq:
    type: object
    required:
      - user_id      
      - timestamp
      - limit
      - isOwner
    properties:
      user_id:
        type: integer
        example: 15
      timestamp:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      limit:
        type: integer
        example: 15
      isOwner:
        type: boolean
        example: true
  PROJECT_RES:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: ting's project
      description:
        type: string
        example: hello world
      isPrivate:
        type: boolean
        example: true
      create_time:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      owner:
        type: string
        example: ting
  ProjectDetailReq:
    type: object
    required:
      - user      
      - project
    properties:
      user:
        type: string
        example: test
      project:
        type: string
        example: test
  ProjectDetailRes:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: ting's project
      description:
        type: string
        example: hello world
      lineId:
        type: string
        example: ajsihdjsfisdi
      isPrivate:
        type: boolean
        example: true
      create_time:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      owner:
        type: string
        example: ting
      ownerAvatarUrl:
        type: string
        example: asdkjahwiuhqihweiqiew
      identity:
        type: integer
      issue_count:
        type: integer
      finish_issue_count:
        type: integer
        example: 1
      member_count:
        type: integer
        example: 1
      tag_count:
        type: integer
        example: 1
      tags:
        type: array
        items: 
          type: string
...
  GetProjectIssueReq:
    type: object
    properties:
      projectId:
        type: integer
        example: 2
      timestamp:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      limit:
        type: integer
        example: 15
  ProjectMemberReq:
    type: object
    required:
      - projectId      
      - username
      - identity
    properties:
      projectId:
        type: integer
        example: 1
      username:
        type: string
        example: test
      identity:
        type: integer
        example: 1
        minimum: 1
  ProjectMemberRes:
    type: object
    properties:
      id:
        type: integer
        example: 2
      username:
        type: string
        example: test
      join_time:
        type: string
        example: 2019-12-05T06:40:43.746+0000
      identity:
        type: integer
        example: 1
      avatarUrl:
        type: string
        example: dawbjkdnawkjndkaw
...
  DeleteProjectMemberReq:
    type: object
    properties:
      projectId:
        type: integer
        example: 1
      userId:
        type: integer
        example: 2
  UpdateProjectMemberReq:
    type: object
    properties:
      projectId:
        type: integer
        example: 1
      userId:
        type: integer
        example: 2
      identity:
        type: integer
        example: 2
  ProjectReportRes:
    type: object
    properties:
      count:
        type: array
        items:
          type: integer
      day:
        type: array
        items:
          type: integer
...